pyproject.toml
[project]
name = "ai-lyrics-finder"
version = "0.3.0"
description = "A LangGraph agent to find, format, and translate song lyrics."
requires-python = ">=3.13"
dependencies = [
    "tavily-python",
    "openai",
    "langgraph",
    "wikipedia",
]

[project.scripts]
find-lyrics-graph = "lyrics_graph:main"

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]
